+++ src/test/java/seedu/address/integration/FindCommandTest.java
@@ -1,0 +1,78 @@
+package seedu.address.integration;
+
+import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
+
+import java.util.List;
+
+import org.junit.Test;
+
+import seedu.address.logic.commands.Command;
+import seedu.address.logic.commands.FindCommand;
+import seedu.address.model.AddressBook;
+import seedu.address.model.person.Person;
+
+public class FindCommandTest extends AddressBookIntegrationTest {
+
+    @Test
+    public void execute_find_invalidArgsFormat() {
+        String expectedMessage = String.format(MESSAGE_INVALID_COMMAND_FORMAT, FindCommand.MESSAGE_USAGE);
+        assertCommandFailure("find ", expectedMessage);
+    }
+
+    @Test
+    public void execute_find_onlyMatchesFullWordsInNames() throws Exception {
+        TestDataHelper helper = new TestDataHelper();
+        Person pTarget1 = helper.generatePersonWithName("bla bla KEY bla");
+        Person pTarget2 = helper.generatePersonWithName("bla KEY bla bceofeia");
+        Person p1 = helper.generatePersonWithName("KE Y");
+        Person p2 = helper.generatePersonWithName("KEYKEYKEY sduauo");
+
+        List<Person> fourPersons = helper.generatePersonList(p1, pTarget1, p2, pTarget2);
+        AddressBook expectedAB = helper.generateAddressBook(fourPersons);
+        List<Person> expectedList = helper.generatePersonList(pTarget1, pTarget2);
+        helper.addToModel(model, fourPersons);
+
+        assertCommandSuccess("find KEY",
+                Command.getMessageForPersonListShownSummary(expectedList.size()),
+                expectedAB,
+                expectedList);
+    }
+
+    @Test
+    public void execute_find_isNotCaseSensitive() throws Exception {
+        TestDataHelper helper = new TestDataHelper();
+        Person p1 = helper.generatePersonWithName("bla bla KEY bla");
+        Person p2 = helper.generatePersonWithName("bla KEY bla bceofeia");
+        Person p3 = helper.generatePersonWithName("key key");
+        Person p4 = helper.generatePersonWithName("KEy sduauo");
+
+        List<Person> fourPersons = helper.generatePersonList(p3, p1, p4, p2);
+        AddressBook expectedAB = helper.generateAddressBook(fourPersons);
+        List<Person> expectedList = fourPersons;
+        helper.addToModel(model, fourPersons);
+
+        assertCommandSuccess("find KEY",
+                Command.getMessageForPersonListShownSummary(expectedList.size()),
+                expectedAB,
+                expectedList);
+    }
+
+    @Test
+    public void execute_find_matchesIfAnyKeywordPresent() throws Exception {
+        TestDataHelper helper = new TestDataHelper();
+        Person pTarget1 = helper.generatePersonWithName("bla bla KEY bla");
+        Person pTarget2 = helper.generatePersonWithName("bla rAnDoM bla bceofeia");
+        Person pTarget3 = helper.generatePersonWithName("key key");
+        Person p1 = helper.generatePersonWithName("sduauo");
+
+        List<Person> fourPersons = helper.generatePersonList(pTarget1, p1, pTarget2, pTarget3);
+        AddressBook expectedAB = helper.generateAddressBook(fourPersons);
+        List<Person> expectedList = helper.generatePersonList(pTarget1, pTarget2, pTarget3);
+        helper.addToModel(model, fourPersons);
+
+        assertCommandSuccess("find key rAnDoM",
+                Command.getMessageForPersonListShownSummary(expectedList.size()),
+                expectedAB,
+                expectedList);
+    }
+}
